Version 1.1
added a wrapper function which makes the insert mode mappings emulate the operator pending mode. i.e, the user does not have to type the LHS fast. Even making mistakes and moving around in the file is allowed before coming back and typing the last character (in this case its chosen to be the tab) expands nicely to the RHS.